feat: stop transforming async functions in default profile
Now that Edge 18 has been released, Edge 14 - the last browser to not
support async functions in the `default` browser set - falls off the
supported list, and we can start shipping native async functions.

While this is kind of a breaking change - in that code that worked will
no longer, it is not a breakage of the _contract_ this package provides
- which is that the `default` target is a moving target that will slowly
drop transformations as more browsers support them. Therefore I consider
this to be a feature change.