feat: ensure rest-spread is always transpiled for browsers that need it.
While babel-preset-env should do this, it also is quite aggressive with
other transforms - for example it adds the destructuring transform for
Edge due to a bug in Edge 15-17. This bug is quite easy to avoid
especially with a lint rule, and so shouldn't really be considered a
problem. However disabling babel-preset-env then introduces another
issue which is that there are still many browsers which need
object-rest-spread transforms! So this patches that issue, whereby any
configurations opt-out of `env` will still get `object-rest-spread`
while the browser list still contains browsers that need it.